+++ src/BasicEnemy.h
@@ -1,12 +1,13 @@
 #pragma once
+
 #include <SDL2/SDL.h>
 
-#include "game_object.h"
-#include "hitbox.h"
+#include "GameObject.h"
+#include "Hitbox.h"
 
-class Missile : public GameObject {
+class BasicEnemy : public GameObject {
 public:
-    Missile(double x_pos, double y_pos, double x_vel, double y_vel);
+    BasicEnemy(double x_pos, double y_pos, double x_vel, double y_vel);
 
     void update();
     void draw(GraphicsHandler *graphics);
@@ -16,6 +17,6 @@ public:
     static void set_texture(SDL_Texture* texture);
 
 private:
-    static SDL_Texture* missile_texture_;
+    static SDL_Texture* texture_;
     Hitbox hitbox_;
 };
