+++ src/CollisionDetector.cpp
@@ -1,5 +1,6 @@
-#include "collision_detector.h"
-#include "hitbox.h"
+#include "GameConstants.h"
+#include "CollisionDetector.h"
+#include "Hitbox.h"
 
 CollisionDetector::CollisionDetector()
 {
@@ -20,4 +21,9 @@ void CollisionDetector::add_projectile(Team::Team team, GameObject& projectile)
 void CollisionDetector::add_game_object(Team::Team team, GameObject& game_object)
 {
     non_projectiles_.push_back(&game_object);
+}
+
+void CollisionDetector::update()
+{
+
 }
