+++ src/Events.cpp
@@ -1,4 +1,4 @@
-#include "events.h"
+#include "Events.h"
 
 void Events::notify_event(SDL_Event e) {
     for(EventsListener *listener : listeners) {
