+++ src/Missile.h
@@ -1,13 +1,12 @@
 #pragma once
-
 #include <SDL2/SDL.h>
 
-#include "game_object.h"
-#include "hitbox.h"
+#include "GameObject.h"
+#include "Hitbox.h"
 
-class BasicEnemy : public GameObject {
+class Missile : public GameObject {
 public:
-    BasicEnemy(double x_pos, double y_pos, double x_vel, double y_vel);
+    Missile(double x_pos, double y_pos, double x_vel, double y_vel);
 
     void update();
     void draw(GraphicsHandler *graphics);
@@ -17,6 +16,6 @@ public:
     static void set_texture(SDL_Texture* texture);
 
 private:
-    static SDL_Texture* texture_;
+    static SDL_Texture* missile_texture_;
     Hitbox hitbox_;
 };
